chore: housekeeping & (re)moved legacy code
