+++ src/__tests__/legacyAPI.spec.ts
@@ -1,4 +1,7 @@
-import { TriggerType, EventType } from './legacyEnums'
+/**
+ * This file is only for legacy reasons and only used in our internal testing
+ */
+import { TriggerType, EventType } from './legacyEnums.spec'
 export interface TimelineTrigger {
 	type: TriggerType,
 	value: number | string
@@ -152,3 +155,6 @@ export interface ResolveOptions {
 	 */
 	limit?: number
 }
+test('tmp', () => {
+	expect(1).toEqual(1)
+})
