+++ src/__tests__/legacyEnums.spec.ts
@@ -28,3 +28,6 @@ export const Enums = {
 	TimelineEventType: EventType,
 	TraceLevel: TraceLevel
 }
+test('tmp', () => {
+	expect(1).toEqual(1)
+})
